64 bit fix
git-svn-id: https://open-ms.svn.sourceforge.net/svnroot/open-ms/OpenMS@2142 6adb6e08-d915-0410-941f-83917bcadc18

+++ source/APPLICATIONS/TOPPBase.C
@@ -415,7 +415,7 @@ namespace OpenMS
 			UInt indent = 0;
 			for(map<String,String>::const_iterator it = subsections_.begin(); it!=subsections_.end(); ++it)
 			{
-				indent = max(it->first.size(),indent);
+				indent = max((UInt)it->first.size(),indent);
 			}
 			indent += 6;
